src/ai-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
python-multipart>=0.0.9
transformers>=4.30.0
nltk>=3.8.0
scipy>=1.10.0
